requirements.txt
FastAPI
SQLModel
pydantic
uvicorn
bcrypt
python-jose
sqlalchemy
databases
email-validator
passlib
dotenv
pydantic
jinja2
python-multipart
fastapi-utils
apscheduler